test: improve test coverage
