+++ test/command/command.test.ts
@@ -1,7 +1,6 @@
 import {expect, fancy} from 'fancy-test'
 // import path = require('path')
-import {Args, Command as Base, Flags} from '../../src'
-import {toCached} from '../../src/util/to-cached'
+import {Command as Base, Flags} from '../../src'
 // import {TestHelpClassConfig} from './helpers/test-help-in-src/src/test-help-plugin'
 
 // const pjson = require('../package.json')
@@ -69,265 +68,6 @@ describe('command', () => {
   })
   .catch(/EEXIT: 0/)
   .it('exits with 0')
-
-  describe('toCached', () => {
-    fancy
-    .do(async () => {
-      class C extends Command {
-        static id = 'foo:bar'
-        static title = 'cmd title'
-        static type = 'mytype'
-        static usage = ['$ usage']
-        static description = 'test command'
-        static aliases = ['alias1', 'alias2']
-        static hidden = true
-        static flags = {
-          flaga: Flags.boolean(),
-          flagb: Flags.string({
-            char: 'b',
-            hidden: true,
-            required: false,
-            description: 'flagb desc',
-            options: ['a', 'b'],
-            default: async () => 'a',
-          }),
-          flagc: Flags.integer({
-            char: 'c',
-            min: 1,
-            max: 10,
-            required: false,
-            description: 'flagc desc',
-            options: ['a', 'b'],
-            default: async context => (context.options.min ?? 1) + 1,
-          }),
-        }
-
-        static args = {
-          arg1: Args.string({
-            description: 'arg1 desc',
-            required: true,
-            hidden: false,
-            options: ['af', 'b'],
-            default: async () => 'a',
-          }),
-        }
-      }
-
-      const c = await toCached(C, undefined, false)
-
-      expect(c).to.deep.equal({
-        id: 'foo:bar',
-        type: 'mytype',
-        hidden: true,
-        pluginName: undefined,
-        pluginAlias: undefined,
-        pluginType: undefined,
-        state: undefined,
-        description: 'test command',
-        aliases: ['alias1', 'alias2'],
-        title: 'cmd title',
-        usage: ['$ usage'],
-        examples: undefined,
-        deprecationOptions: undefined,
-        deprecateAliases: undefined,
-        summary: undefined,
-        strict: true,
-        enableJsonFlag: false,
-        hasDynamicHelp: false,
-        flags: {
-          flaga: {
-            aliases: undefined,
-            char: undefined,
-            charAliases: undefined,
-            description: undefined,
-            dependsOn: undefined,
-            deprecateAliases: undefined,
-            deprecated: undefined,
-            exclusive: undefined,
-            helpGroup: undefined,
-            helpLabel: undefined,
-            summary: undefined,
-            name: 'flaga',
-            hidden: undefined,
-            required: undefined,
-            relationships: undefined,
-            allowNo: false,
-            type: 'boolean',
-            delimiter: undefined,
-            noCacheDefault: undefined,
-          },
-          flagb: {
-            aliases: undefined,
-            char: 'b',
-            charAliases: undefined,
-            description: 'flagb desc',
-            dependsOn: undefined,
-            deprecateAliases: undefined,
-            deprecated: undefined,
-            exclusive: undefined,
-            helpGroup: undefined,
-            helpLabel: undefined,
-            summary: undefined,
-            name: 'flagb',
-            hidden: true,
-            required: false,
-            multiple: false,
-            relationships: undefined,
-            type: 'option',
-            helpValue: undefined,
-            default: 'a',
-            options: ['a', 'b'],
-            delimiter: undefined,
-            noCacheDefault: undefined,
-          },
-          flagc: {
-            aliases: undefined,
-            char: 'c',
-            charAliases: undefined,
-            default: 2,
-            delimiter: undefined,
-            dependsOn: undefined,
-            deprecateAliases: undefined,
-            deprecated: undefined,
-            description: 'flagc desc',
-            exclusive: undefined,
-            helpGroup: undefined,
-            helpLabel: undefined,
-            helpValue: undefined,
-            hidden: undefined,
-            multiple: false,
-            name: 'flagc',
-            options: [
-              'a',
-              'b',
-            ],
-            relationships: undefined,
-            required: false,
-            summary: undefined,
-            type: 'option',
-            noCacheDefault: undefined,
-          },
-
-        },
-        args: {
-          arg1: {
-            description: 'arg1 desc',
-            name: 'arg1',
-            hidden: false,
-            required: true,
-            options: ['af', 'b'],
-            default: 'a',
-            noCacheDefault: undefined,
-          },
-        },
-      })
-    })
-    .it('converts to cached with everything set')
-
-    fancy
-    .do(async () => {
-      class Base extends Command {
-        public static enableJsonFlag = true
-        public static baseFlags = {
-          parentFlag: Flags.boolean(),
-        }
-      }
-
-      class Child extends Base {
-        static flags = {
-          childFlag: Flags.boolean(),
-        }
-      }
-
-      const cached = await toCached(Child, undefined, false)
-
-      expect(cached).to.deep.equal({
-        id: 'command',
-        summary: undefined,
-        description: 'test command',
-        strict: true,
-        usage: undefined,
-        pluginName: undefined,
-        pluginAlias: undefined,
-        pluginType: undefined,
-        hidden: undefined,
-        state: undefined,
-        aliases: [],
-        examples: undefined,
-        deprecationOptions: undefined,
-        deprecateAliases: undefined,
-        flags: {
-          json: {
-            name: 'json',
-            type: 'boolean',
-            char: undefined,
-            summary: undefined,
-            description: 'Format output as json.',
-            hidden: undefined,
-            required: undefined,
-            helpLabel: undefined,
-            helpGroup: 'GLOBAL',
-            allowNo: false,
-            dependsOn: undefined,
-            relationships: undefined,
-            exclusive: undefined,
-            deprecated: undefined,
-            deprecateAliases: undefined,
-            aliases: undefined,
-            charAliases: undefined,
-            delimiter: undefined,
-            noCacheDefault: undefined,
-          },
-          childFlag: {
-            name: 'childFlag',
-            type: 'boolean',
-            char: undefined,
-            summary: undefined,
-            description: undefined,
-            hidden: undefined,
-            required: undefined,
-            helpLabel: undefined,
-            helpGroup: undefined,
-            allowNo: false,
-            dependsOn: undefined,
-            relationships: undefined,
-            exclusive: undefined,
-            deprecated: undefined,
-            deprecateAliases: undefined,
-            aliases: undefined,
-            charAliases: undefined,
-            delimiter: undefined,
-            noCacheDefault: undefined,
-          },
-          parentFlag: {
-            name: 'parentFlag',
-            type: 'boolean',
-            char: undefined,
-            summary: undefined,
-            description: undefined,
-            hidden: undefined,
-            required: undefined,
-            helpLabel: undefined,
-            helpGroup: undefined,
-            allowNo: false,
-            dependsOn: undefined,
-            relationships: undefined,
-            exclusive: undefined,
-            deprecated: undefined,
-            deprecateAliases: undefined,
-            aliases: undefined,
-            charAliases: undefined,
-            delimiter: undefined,
-            noCacheDefault: undefined,
-          },
-        },
-        args: {},
-        hasDynamicHelp: false,
-        enableJsonFlag: true,
-      })
-    })
-    .it('converts to cached with multiple Command classes in inheritance chain')
-  })
 
   describe('parse', () => {
     fancy
